ReplicatedPG: merge apply_ctx_scrub_stats into apply_stats
The only way the two aren't called essentially atomically
is when the ECBackend needs to delay a write for rmw.  In
such a case, we actually want to delay the scrub_cstat
application as well since we otherwise, we'd have to flush
the backend before settling up at the end of scrub.

Perhaps this should be a separate function to distinguish
between the replica+primary info update and the primary
scrub_cstat update, but that seems more complicated than
it's really worth.

Fixes: http://tracker.ceph.com/issues/17999
Signed-off-by: Samuel Just <sjust@redhat.com>