build/ops: long ceph-osd@.service pause between failures
Instead of the default 100ms pause before trying to restart an OSD, wait
5 minutes instead. There is no scenario in which restarting an OSD
almost immediately after it failed would get a better result.

It is however possible that a failure to start is due to a race with
another systemd unit at boot time. For instance if ceph-disk@.service is
delayed, it may start after the OSD that needs it. A long pause may
give the racing service enough time to complete and the next attempt to
start the OSD may succeed.

This is not a sound alternative to resolve a race, it only makes the
OSD boot process less sensitive.

Refs: http://tracker.ceph.com/issues/17889

Signed-off-by: Loic Dachary <loic@dachary.org>